Update README.md
Added material to bring this repository in line with (A)GPL expectations around licensing, and clearly visible notices as per https://github.com/th3j35t3r/CounterSocial/issues/1 . We really didn't need to, but in the interests of goodwill, we did so anyway. Because we're actually decent like that.